Extract a shared rdx integration test helper
